rcbl-backend/pyproject.toml
[project]
name = "rcbl-backend"
version = "0.1.0"
description = "RCBL receivables backend — invoice upload + OCR module"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "sqlalchemy[asyncio]>=2.0.47",
    "asyncpg>=0.31.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=46.0.5",
    "redis[hiredis]>=5.3.1",
    "arq>=0.27.0",
    "slowapi>=0.1.9",
    "structlog>=25.5.0",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "httpx>=0.28.1",
    "aioboto3>=15.5.0",
    "python-slugify>=8.0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
    "mypy>=1.19.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
    "mypy>=1.19.1",
    "aiosqlite>=0.22.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
